WORKING!!! UPDATE: Removed Some (Not All) Dead Code; Move AppForm to Route/Form
